a refactor and matching table styling to figma
