+++ frontend/src/components/CostTable/index.css
@@ -1,0 +1,37 @@
+.table {
+  margin-top: 30px;
+}
+
+/* Antd global styling messing with our desired styling
+So try and beat it with CSS specificity */
+.ant-table-container,
+.ant-table-container table > thead > tr:first-child > th {
+  background-color: #d9d9d9;
+  border: 1px solid #d9d9d9;
+}
+.ant-table-container,
+.ant-table-container table > thead > tr:first-child th:first-child {
+  border-top-left-radius: 10px;
+  border-color: #d9d9d9;
+}
+.ant-table-container,
+.ant-table-container table > thead > tr:first-child th:last-child {
+  border-top-right-radius: 10px;
+  border-color: #d9d9d9;
+}
+.ant-table-container,
+.ant-table-container table > tbody > tr {
+  background-color: #fff;
+}
+.ant-table-container,
+.ant-table-container table > tbody > tr td {
+  border: 1px solid #d9d9d9;
+}
+.ant-table-container,
+.ant-table-container table > tbody > tr:last-child td:first-child {
+  border-bottom-left-radius: 10px;
+}
+.ant-table-container,
+.ant-table-container table > tbody > tr:last-child td:last-child {
+  border-bottom-right-radius: 10px;
+}
