--- frontend/src/screens/billing-management/CostTable/index.css
@@ -1,54 +1,0 @@
-
-.table{
-    margin: 0px;
-    padding: 0px;
-    height: 100%;
-    width: 100%;
-    margin-top: 30px;
-    margin-bottom: 30px;
-    border: 1px solid #3f3d3d;
-    border-radius: 15px;
-    overflow: hidden;
-    /* box-shadow: 0 0 20px rgba(0, 0, 0, 0.15); */
-    padding: 5px;
-    background-color: white;
-}
-
-.table td{
-    padding: 5px;
-    text-align: center;
-    border: 0.1px solid #e6dede;
-    line-height: 1.5;
-    height: 30px;
-}
-
-.table th{
-    text-align: center;
-    padding: 5px;
-    font-size: 18px;
-    color: #878484;
-    align-items: center;
-    justify-content: center;
-    background-color: #878484;
-    border: 0.1px solid #e6dede;
-}
-
-.table tr{
-    border: 1px solid #3f3d3d;
-    border-radius: 15px;
-    overflow: hidden;
-    /* box-shadow: 0 0 20px rgba(0, 0, 0, 0.15); */
-    padding: 5px;
-    text-align: center;
-    align-items: center;
-    justify-content: center;
-    background-color: white;
-}
-
-.table tr:nth-child(even){
-    background-color: #f2f2f2;
-}
-
-.table tr:nth-child(odd){
-    background-color: #ffffff;
-}
